+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 This file holds the Changelog of the project and all relevant changes.
 
-## 1.0.0
+## 1.1.0
+### Added
+- Implemented a console printer for formatted output per telegram
+- Implemented TelegramPlayer for replaying telegrams from files
+- First base information for GSM status
 
+## 1.0.0
 Initial release including support for live parsing and parsing of stored files.